[Rotate logos] Fixed refreshing logos from admin panel

+++ apps/rotatelogos/views.py
@@ -14,13 +14,17 @@ def index(request):
 	return render_to_response('rotatelogos/index.html', {"img_url": img_url}, context_instance=RequestContext(request))
 
 def refresha(request):
+	status = ""
 	# Reads in logos from the folder, then create or remove thumbnails
 	# TODO: User rights instead of just authentication...
 	if not request.user.is_authenticated():
-		return render_to_response('rotatelogos/failed.html', {}, context_instance=RequestContext(request))
+		status = "Refresh failed, not authenticated \n"
+		return render_to_response('rotatelogos/status.html', {"status":status}, context_instance=RequestContext(request))
 	
-	image_path = settings.MEDIA_ROOT + '/' + settings.FILEBROWSER_DIRECTORY + '/rotatelogos/' + '*.jpg'
-	thumbs_path = settings.MEDIA_ROOT + '/' + settings.FILEBROWSER_DIRECTORY + '/rotatelogos/.thumbs/' + '*.gif|*.jpg|*.png|*.jpeg'
+	image_formats = ["gif","png","tif","jpeg","jpg"]
+
+	image_path = settings.MEDIA_ROOT + '/' + settings.FILEBROWSER_DIRECTORY + '/rotatelogos/*'
+	thumbs_path = settings.MEDIA_ROOT + '/' + settings.FILEBROWSER_DIRECTORY + '/rotatelogos/.thumbs/*'
 
 	image_path = image_path.replace("//","/")
 	thumbs_path = thumbs_path.replace("//","/")
@@ -30,10 +34,9 @@ def refresha(request):
 	image_list_names = {}
 	thumbs_list_names = {}
 
-	print image_path
-	print thumbs_path
-	print image_list
+	print image_list 
 	print thumbs_list
+	
 
 	for x in thumbs_list:
 		thumbs_list_names[x] = x.split("/")[-1]
@@ -41,8 +44,11 @@ def refresha(request):
 	for x in image_list:
 		image_list_names[x] = x.split("/")[-1]
 
+	print image_list_names 
+	print thumbs_list_names 
+
 	for x,y in image_list_names.iteritems():
-		if y not in thumbs_list_names:
+		if y not in thumbs_list_names.values() and y.split(".")[-1].lower() in image_formats:
 			#Make thumbnail
 			outfile = settings.MEDIA_ROOT + \
 							'/' + settings.FILEBROWSER_DIRECTORY + \
@@ -52,12 +58,14 @@ def refresha(request):
 			im = Image.open(infile)
 			im.thumbnail((300,90), Image.ANTIALIAS)
 			im.save(outfile, "JPEG")
-			print "made shit"
+			status += "Added " + y + " thumbnail \n"
 
 
 	for x,y in thumbs_list_names.iteritems():
-		if y not in image_list_names:
+		if y not in image_list_names.values() and y.split(".")[-1].lower() in image_formats:
 			#Remove thumbnail
 			os.remove(x)
+			status += "Removed thumbnail " + y +"\n"
 
-	return render_to_response('rotatelogos/created.html', {}, context_instance=RequestContext(request))
+	status += "All done.. :)"
+	return render_to_response('rotatelogos/status.html', {"status":status}, context_instance=RequestContext(request))
